refactor: move the backup endpoints out of the closure
Fourth group of #669. BackupList, BackupCreate, BackupDownload, BackupRestore
and BackupDelete move into modules/api/resources_backup.py, built by
create_backup_resources(api, models, ctx).

_validate_backup_filename moves with them — it is used only by these five
endpoints, and leaving it in resources.py would have made this module import
the one that imports it. resources.py re-exports it so the existing tests keep
their import path.

Closure complexity 485 -> 439; resources.py 3616 -> 3262 lines.

Which classes were safe to move together was decided by asking the AST which
names each one captures from the enclosing scope, rather than by reading:
all five need only auth, settings, file_ops, audit and the manager mapping.
The same analysis says 18 of the 36 remaining classes are already buildable
from the context alone, and the rest need only the thin helper wrappers the
prologue defines, so the remaining groups are mechanical.

The move itself is not, quite. The rewrite of captured names to ctx.* went
through a regex that did not respect string boundaries, and turned
managers.get('file_ops') into managers.get('ctx.file_ops') — BackupDelete then
answered 503 to every request. Every build-only test still passed, because
none of them call the method. The full suite caught it.

test_api_group_modules_are_standalone.py therefore gains a case that calls
BackupDelete against a real file and asserts it did the work, so a wrong
lookup key fails at the group level instead of surviving to the suite.
Verified by reintroducing that exact corruption: the new case fails, the
build-only cases stay green.
